Combined changes from last 8 commits
