+++ packages/dressed/jsr.json
@@ -5,10 +5,20 @@
   "exports": {
     ".": "./src/index.ts",
     "./server": "./src/server/index.ts",
-    "./utils": "./src/utils/index.ts"
+    "./utils": "./src/utils/index.ts",
+    "./handler": "./src/handler/index.ts"
   },
   "publish": {
-    "include": ["README.md", "src"],
-    "exclude": ["dist", "!src/resources/generated.resources.ts", "src/resources/make", "**/*.test.ts"]
-  }
+    "include": [
+      "README.md",
+      "src"
+    ],
+    "exclude": [
+      "dist",
+      "!src/resources/generated.resources.ts",
+      "src/resources/make",
+      "**/*.test.ts"
+    ]
+  },
+  "sideEffects": false
 }
