+++ www/content/guide/deploying/vercel.md
@@ -24,7 +24,7 @@ Deploying is the last step in building your bot, it will be where Discord can se
 3. ```ts title="api / bot.ts"
    // @ts-ignore Generated after build
    import { commands, components, events } from "../.dressed/index.js";
-   import { handleRequest } from "dressed/server";
+   import { handleRequest } from "dressed/handler";
 
    export const POST = (req: Request) => handleRequest(req, commands, components, events);
    ```
@@ -38,7 +38,7 @@ If you’re using Next.js, you don’t need a `vercel.json` file. The Next.js fr
 ```ts title="app / api / bot / route.ts"
 // @ts-ignore Generated after build
 import { commands, components, events } from "../../../.dressed";
-import { handleRequest } from "dressed/server";
+import { handleRequest } from "dressed/handler";
 
 export const POST = (req: Request) => handleRequest(req, commands, components, events);
 ```
